Verify launch blockers on real PostgreSQL: RLS suite, P&L honesty, numeric parity
Backend suite: 197 passed / 0 failed on PostgreSQL 16.2 with FORCE RLS
(enabled via local pgserver; previously SQLite-only in this environment).
SQLite fallback also green: 188 passed, 9 PG-only skips.

Production fixes found by running against real Postgres:
- providers connect: drop post-commit db.refresh(); commit() ends the
  transaction and drops SET LOCAL app.org_id, so the refresh SELECT came
  back empty under FORCE RLS and raised InvalidRequestError.
- anomaly acknowledge: query tenant names BEFORE db.commit() for the same
  reason; afterwards the lookup silently returned zero rows.
- investigate service: coerce the two pct deltas to float. On Postgres,
  sum() over integer columns returns Decimal while SQLite returns int,
  so round() produced Decimal and broke the schema-declared float
  contract (Pydantic masked it at the API boundary).

RLS test expansion (tests/test_rls.py, Postgres-only):
- per-table cross-org isolation for projects, tenants, usage events,
  cost anomalies, recommendations, provider credentials, and API keys
- billing-shaped columns on the organization row do not leak across orgs
- update/delete/insert blocked, not just reads

P&L honesty (unpriced costs never silently $0):
- _tenant_costs counts NULL-cost events per tenant; schemas expose
  unpriced_events; P&L cards render 'Excludes N unpriced event(s) -
  cost unknown, not $0.'

Test-only: pin SET LOCAL app.org_id on standalone committed sessions
and re-pin after commit() where a service call follows.

Frontend: tsc --noEmit, ESLint, and production build all green.

+++ app/backend/app/services/investigate.py
@@ -108,8 +108,11 @@ def _volume_vs_tokens(db: Session, org_id: uuid.UUID, project_id: uuid.UUID,
                       - volume_effect - token_effect).quantize(Decimal("0.01"))
 
     return {
-        "requests_delta_pct": round(requests_delta, 1),
-        "avg_tokens_per_request_delta_pct": round(tpr_delta, 1),
+        # Pct deltas are display fields declared as float in the schemas;
+        # coerce explicitly: on Postgres, sum() over integer columns comes
+        # back as Decimal while SQLite returns int.
+        "requests_delta_pct": float(round(requests_delta, 1)),
+        "avg_tokens_per_request_delta_pct": float(round(tpr_delta, 1)),
         "window_note": "last 7 days vs prior 23 days",
         "volume_effect_usd": volume_effect,
         "token_intensity_effect_usd": token_effect,
